chore(TCOMP-216): move reader test use cases to test folder
